Fully revert write barrier changes
